+++ autodarts-caller.py
@@ -143,6 +143,7 @@ CALLER_PROFILES = {
     'en-US-Joey-Male': ('https://add.arnes-design.de/ADC/en-US-Joey-Male-v4.zip', 4),
     'en-US-Joanna-Female': ('https://add.arnes-design.de/ADC/en-US-Joanna-Female-v4.zip', 4),
     'en-US-Gregory-Male': ('https://add.arnes-design.de/ADC/en-US-Gregory-Male.zip', 1),
+    'en-US-Matthew-Male': ('https://add.arnes-design.de/ADC/en-US-Matthew-Male.zip', 1),
 
     
     # 'TODONAME': ('TODOLINK', TODOVERSION),  
@@ -233,7 +234,10 @@ def check_paths(main_directory, audio_media_path, audio_media_path_shared, black
 
     try:
         main_directory = os.path.normpath(os.path.dirname(os.path.realpath(main_directory)))
+        os.chdir(main_directory)  # Set the working directory to main_directory
+
         audio_media_path = os.path.normpath(audio_media_path)
+        
         if audio_media_path_shared != DEFAULT_EMPTY_PATH:
             audio_media_path_shared = os.path.normpath(audio_media_path_shared)
         if blacklist_path != DEFAULT_EMPTY_PATH:
@@ -258,6 +262,12 @@ def check_paths(main_directory, audio_media_path, audio_media_path_shared, black
 
     except Exception as e:
         errors = f'Path validation failed: {e}'
+
+    if errors != None:
+        ppi("main_directory: " + main_directory)
+        ppi("audio_media_path: " + str(audio_media_path))
+        ppi("audio_media_path_shared: " + str(audio_media_path_shared))
+        ppi("blacklist_path: " + str(blacklist_path))
 
     return errors
 
@@ -476,7 +486,7 @@ def ban_caller(only_change):
     
 
 
-def load_callers_banned():
+def load_callers_banned(preview = False):
     global caller_profiles_banned
     caller_profiles_banned = []
     
@@ -487,6 +497,12 @@ def load_callers_banned():
     if os.path.exists(path_to_callers_banned_file):
         with open(path_to_callers_banned_file, 'r') as bcf:
             caller_profiles_banned = list(set(line.strip() for line in bcf))
+            if preview:
+                banned_info = f"Banned voice-packs: {len(caller_profiles_banned)} [ - "
+                for cpb in caller_profiles_banned:
+                    banned_info += cpb + " - "
+                banned_info += "]"
+                ppi(banned_info)
     else:
         with open(path_to_callers_banned_file, 'x'):
             ppi(f"'{path_to_callers_banned_file}' created successfully.")
@@ -671,7 +687,7 @@ def setup_caller():
         # ppi(caller[1])
         caller = caller[1]
 
-        # TODO: woanders
+
         # files = []
         # for key, value in caller.items():
         #     for sound_file in value:
@@ -681,15 +697,16 @@ def setup_caller():
         #     "caller": caller_title_without_version,
         #     "files": files
         # }
-        # broadcast(get_event)
+        # if server != None:
+        #   broadcast(get_event)
 
-        # TODO: "caller-list": callers
         welcome_event = {
             "event": "welcome",
             "caller": caller_title_without_version,
             "caller-version": caller_title,   
         }
-        broadcast(welcome_event)
+        if server != None:
+            broadcast(welcome_event)
 
 
 def play_sound(sound, wait_for_last, volume_mult):
@@ -2182,14 +2199,6 @@ def on_error_autodarts(ws, error):
 def on_open_client(client, server):
     ppi('NEW CLIENT CONNECTED: ' + str(client))
 
-    # TODO: "caller-list": callers
-    welcome_event = {
-        "event": "welcome",
-        "caller": caller_title_without_version,
-        "caller-version": caller_title,   
-    }
-    unicast(client, welcome_event)
-
 def on_message_client(client, server, message):
     def process(*args):
         try:
@@ -2267,7 +2276,16 @@ def on_message_client(client, server, message):
             #         "files": files
             #     }
             #     unicast(client, get_event)
-            
+
+            elif message.startswith('hello'):
+                # TODO: "caller-list": callers
+                welcome_event = {
+                    "event": "welcome",
+                    "caller": caller_title_without_version,
+                    "caller-version": caller_title,   
+                }
+                unicast(client, welcome_event)
+
             # else try to read json
             else: 
                 messageJson = json.loads(message)
@@ -2278,7 +2296,8 @@ def on_message_client(client, server, message):
                     new = []
                     count_exists = 0
                     count_new = 0
-                    for key, value in caller.items():
+                    caller_copied = caller.copy()
+                    for key, value in caller_copied.items():
                         for sound_file in value:
                             # care encoding!
                             base_name = os.path.basename(sound_file)
@@ -2303,7 +2322,7 @@ def on_message_client(client, server, message):
 
           
         except Exception as e:
-            ppe('WS-Client-Message failed: ', e)
+            ppe('WS-Client-Message failed. Message: ' + message, e)
 
     t = threading.Thread(target=process).start()
 
@@ -2316,7 +2335,7 @@ def broadcast(data):
         server.send_message_to_all(json.dumps(data, indent=2).encode('utf-8'))
     t = threading.Thread(target=process)
     t.start()
-    t.join()
+    # t.join()
 
 def unicast(client, data, dump=True):
     def process(*args):
@@ -2602,8 +2621,9 @@ if __name__ == "__main__":
             mixer.pre_init(MIXER_FREQUENCY, MIXER_SIZE, MIXER_CHANNELS, MIXER_BUFFERSIZE)
             mixer.init()
         except Exception as e:
-            ppe("Failed to initialize audio device! Make sure the target device is connected and configured as os default. A device connected by HDMI can cause problems; use standard audio-jack instead.", e)
-            sys.exit()  
+            WEB = 1
+            ppe("Failed to initialize audio device! Make sure the target device is connected and configured as os default. Fallback to web-caller", e)
+            # sys.exit()  
 
     path_status = check_paths(__file__, AUDIO_MEDIA_PATH, AUDIO_MEDIA_PATH_SHARED, BLACKLIST_PATH)
     if path_status is not None: 
@@ -2620,7 +2640,7 @@ if __name__ == "__main__":
             ppe("Background-Muter failed!", e)
 
     try:
-        load_callers_banned()
+        load_callers_banned(preview = True)
         download_callers()
     except Exception as e:
         ppe("Voice-pack fetching failed!", e)