+++ static/js/common.js
@@ -21,15 +21,27 @@ function setupWebSocketConnection(host, port, onOpenHandle, onCloseHandle, onMes
         onOpenHandle(socket);
     };
 
-    socket.onclose = function() {
+    socket.onclose = function(event) {
         console.log("Socket: disconnected!");
         onCloseHandle(socket);
+
+        if (event.wasClean) {
+            console.log('Verbindung geschlossen sauber.');
+        } else {
+            console.log('Verbindung unerwartet geschlossen.');
+        }
+        console.log('Code: ' + event.code + ', Grund: ' + event.reason);
+
         setTimeout(() => setupWebSocketConnection(host, port, onOpenHandle, onCloseHandle, onMessageHandle), 1000);
     };
 
     socket.onmessage = function(event) {
         let data = JSON.parse(event.data);
         onMessageHandle(socket, data); 
+    };
+
+    socket.onerror = function(error) {
+        console.error('WebSocket-Fehler aufgetreten: ', error);
     };
 
     return socket;