+++ static/js/common.js
@@ -11,8 +11,6 @@ function getRandomInterval(min, max) {
 }
 
 
-
-
 function setupWebSocketConnection(host, port, onOpenHandle, onCloseHandle, onMessageHandle, maxReconnectAttempts = 3, reconnectInterval = 1000) {
     let socket;
 
@@ -28,10 +26,11 @@ function setupWebSocketConnection(host, port, onOpenHandle, onCloseHandle, onMes
             console.log("Socket: disconnected!");
             onCloseHandle(socket, event);
 
-            if (event.code !== 1000 && maxReconnectAttempts > 0) {
+            // && maxReconnectAttempts > 0
+            if (event.code !== 1000) {
                 console.log(`Reconnecting in ${reconnectInterval / 1000} seconds...`);
                 setTimeout(connect, reconnectInterval);
-                maxReconnectAttempts--;
+                // maxReconnectAttempts--;
             }
         };
 
@@ -48,38 +47,8 @@ function setupWebSocketConnection(host, port, onOpenHandle, onCloseHandle, onMes
             console.error('WebSocket error occurred:', error);
         };
     }
-
     connect();
 
     return socket;
 }
-
-
-// function setupWebSocketConnection(host, port, onOpenHandle, onCloseHandle, onMessageHandle) {
-//     const socket = new WebSocket(`ws://${host}:${port}`);
-
-//     socket.onopen = function() {
-//         console.log("Socket: connected!");
-//         onOpenHandle(socket);
-//     };
-
-//     socket.onclose = function(event) {
-//         console.log("Socket: disconnected!");
-//         onCloseHandle(socket);
-
-//         console.log('Code: ' + event.code + ', Reason: ' + event.reason + ', Clean: ' + event.wasClean);
-//         setTimeout(() => setupWebSocketConnection(host, port, onOpenHandle, onCloseHandle, onMessageHandle), 1000);
-//     };
-
-//     socket.onmessage = function(event) {
-//         let data = JSON.parse(event.data);
-//         onMessageHandle(socket, data); 
-//     };
-
-//     socket.onerror = function(error) {
-//         console.error('WebSocket-Fehler aufgetreten: ', error);
-//     };
-
-//     // return socket;
-// }