+++ autodarts-caller.py
@@ -48,7 +48,7 @@ main_directory = os.path.dirname(os.path.realpath(__file__))
 parent_directory = os.path.dirname(main_directory)
 
 
-VERSION = '2.7.4'
+VERSION = '2.7.5'
 
 
 DEFAULT_EMPTY_PATH = ''