+++ CHANGELOG.md
@@ -4,8 +4,7 @@
 - add voice-pack 'en-US-Matthew-Male'
 - preview banned voice-packs
 - add fallback for unreachable audio device
-- add lobby-event broadcasts
-- fix invalid -M and -MS
+- fix invalid -M (-MS)
 
 
 ## 2.7.3