+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - add voice-pack 'en-US-Matthew-Male'
 - preview banned voice-packs
 - add fallback for unreachable audio device
+- prevent double execution
 - fix invalid -M (-MS)