+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 - add voice-pack 'en-US-Matthew-Male'
 - preview banned voice-packs
 - add fallback for unreachable audio device
-- fix invalid -M (-MS)
+- prevent app from start twice
 
 
 ## 2.7.3